src/requirements.txt
appdirs==1.4.0
astroid==1.4.9
colorama==0.3.7
Django==1.10.3
django-crispy-forms==1.6.1
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
drfdocs==0.0.11
isort==4.2.5
lazy-object-proxy==1.2.2
Markdown==2.6.7
markdown2==2.3.3
mccabe==0.6.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
PyJWT==1.4.2
pylint==1.6.5
pyparsing==2.1.10
six==1.10.0
wrapt==1.10.8
